Working on actions

+++ test_pipeline.sh
@@ -12,6 +12,22 @@ make WorkloadStatistics SimpleExecutor hyriseServer
 ./hyriseServer --benchmark_data=TPC-H:10 &
 server_pid=$!
 
+cat > commands.sql << EOF
+INSERT INTO meta_plugins(name) VALUES ('rel/libWorkloadHandler.so');
+select count(*) from meta_benchmark_items;
+EOF
 
+cat > TPCH_Q01.exp << EOF
+#!/usr/bin/expect
+spawn psql -h localhost -p 5432 -f commands.sql
+expect "22"
+EOF
+
+chmod +x TPCH_Q01.exp
+./TPCH_Q01.exp
+ret_code=$?
+if [ $ret_code != 0 ]; then
+	exit $ret_code
+fi
 
 kill -9 $server_pid
